feat : artwork dto 생성

+++ Arti-Backend/src/main/java/com/d106/arti/artwork/repository/ArtworkRepository.java
@@ -5,5 +5,6 @@ import com.d106.arti.artwork.domain.Artwork;
 import com.d106.arti.artwork.domain.NormalArtWork;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ArtworkRepository extends JpaRepository<NormalArtWork, Long> {
+public interface ArtworkRepository extends JpaRepository<NormalArtWork, Integer> {
+
 }
